Further optimization work, SpoutChunk and such is utterly broken right now :D. Signed-off-by: Chris Sanders <zidane@outlook.com>
