+++ src/main/java/org/spout/engine/entity/EntityManager.java
@@ -30,37 +30,32 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import org.spout.api.Spout;
 import org.spout.api.datatable.GenericDatatableMap;
 import org.spout.api.entity.Controller;
 import org.spout.api.entity.Entity;
 import org.spout.api.entity.Player;
 import org.spout.api.entity.controller.BlockController;
 import org.spout.api.entity.controller.PlayerController;
-import org.spout.api.event.entity.EntityDespawnEvent;
-import org.spout.api.event.entity.EntitySpawnEvent;
 import org.spout.api.math.Vector3;
+import org.spout.api.protocol.NetworkSynchronizer;
 import org.spout.api.util.StringMap;
 
 import org.spout.engine.util.thread.snapshotable.SnapshotManager;
 import org.spout.engine.util.thread.snapshotable.SnapshotableArrayList;
 import org.spout.engine.util.thread.snapshotable.SnapshotableHashMap;
-import org.spout.engine.util.thread.snapshotable.SnapshotableHashSet;
 import org.spout.engine.world.SpoutChunk;
 import org.spout.engine.world.SpoutRegion;
 
 /**
  * A class which manages all of the entities within a world.
  */
-public class EntityManager implements Iterable<SpoutEntity> {
+public class EntityManager {
 	/**
 	 * The snapshot manager
 	 */
@@ -99,9 +94,9 @@ public class EntityManager implements Iterable<SpoutEntity> {
 	public List<SpoutEntity> getAll(Class<? extends Controller> type) {
 		SnapshotableArrayList<SpoutEntity> entities = groupedEntities.get(type);
 		if (entities == null) {
-			return null;
+			return Collections.emptyList();
 		}
-		return entities.get();
+		return Collections.unmodifiableList(entities.get());
 	}
 
 	/**
@@ -112,9 +107,17 @@ public class EntityManager implements Iterable<SpoutEntity> {
 	public List<SpoutEntity> getAll() {
 		Collection<SpoutEntity> all = entities.get().values();
 		if (all == null) {
-			return null;
+			return Collections.emptyList();
 		}
-		return new ArrayList<SpoutEntity>(all);
+		return Collections.unmodifiableList(new ArrayList<SpoutEntity>(all));
+	}
+
+	public List<SpoutEntity> getAllLive() {
+		Collection<SpoutEntity> all = entities.getLive().values();
+		if (all == null) {
+			return Collections.emptyList();
+		}
+		return Collections.unmodifiableList(new ArrayList<SpoutEntity>(all));
 	}
 
 	/**
@@ -223,12 +226,11 @@ public class EntityManager implements Iterable<SpoutEntity> {
 	 * @return players managed by this entity manager
 	 */
 	public List<Player> getPlayers() {
-		return new ArrayList<Player>(players.get());
-	}
-
-	@Override
-	public Iterator<SpoutEntity> iterator() {
-		return entities.get().values().iterator();
+		List<SpoutPlayer> thePlayers = players.get();
+		if (thePlayers == null) {
+			return Collections.emptyList();
+		}
+		return Collections.unmodifiableList(new ArrayList<Player>(thePlayers));
 	}
 
 	public void finalizeRun() {
@@ -272,7 +274,7 @@ public class EntityManager implements Iterable<SpoutEntity> {
 	 */
 	public void copyAllSnapshots() {
 		for (SpoutEntity e : entities.get().values()) {
-			e.copyToSnapshot();
+			e.copySnapshot();
 		}
 		snapshotManager.copyAllSnapshots();
 	}
@@ -296,5 +298,138 @@ public class EntityManager implements Iterable<SpoutEntity> {
 
 	public Map<Vector3, Entity> getBlockEntities() {
 		return Collections.unmodifiableMap(blockEntities);
+	}
+
+	public void syncEntities() {
+		Map<Entity, Integer> observerSnapshot = observers.get();
+		Map<Entity, Integer> observerLive = observers.getLive();
+
+		List<Entity> entitiesSnapshot = entities.get();
+
+		// Changed means entered/left the chunk
+		List<Entity> changedEntities = entities.getDirtyList();
+		List<Entity> changedObservers = observers.getDirtyList();
+
+		if (entitiesSnapshot.size() > 0) {
+			for (Entity p : changedObservers) {
+				Integer playerDistanceOld = observerSnapshot.get(p);
+				if (playerDistanceOld == null) {
+					playerDistanceOld = Integer.MAX_VALUE;
+				}
+				Integer playerDistanceNew = observerLive.get(p);
+				if (playerDistanceNew == null) {
+					playerDistanceNew = Integer.MAX_VALUE;
+				}
+				// Player Network sync
+				if (p instanceof Player) {
+					Player player = (Player) p;
+
+					NetworkSynchronizer n = player.getNetworkSynchronizer();
+					for (Entity e : entitiesSnapshot) {
+						if (player.equals(e)) {
+							continue;
+						}
+						int entityViewDistanceOld = ((SpoutEntity) e).getPrevViewDistance();
+						int entityViewDistanceNew = e.getViewDistance();
+
+						if (playerDistanceOld <= entityViewDistanceOld && playerDistanceNew > entityViewDistanceNew) {
+							n.destroyEntity(e);
+						} else if (playerDistanceNew <= entityViewDistanceNew && playerDistanceOld > entityViewDistanceOld) {
+							n.spawnEntity(e);
+						}
+					}
+				}
+			}
+		}
+
+		for (Entity e : changedEntities) {
+			SpoutChunk oldChunk = (SpoutChunk) e.getChunk();
+			if (((SpoutEntity) e).justSpawned()) {
+				oldChunk = null;
+			}
+			SpoutChunk newChunk = (SpoutChunk) ((SpoutEntity) e).getChunkLive();
+			if (!(oldChunk != null && oldChunk.equals(this)) && !((SpoutEntity) e).justSpawned()) {
+				continue;
+			}
+			for (Entity p : observerLive.keySet()) {
+				if (p == null || p.equals(e)) {
+					continue;
+				}
+				if (p.getController() instanceof PlayerController) {
+					Integer playerDistanceOld;
+					if (oldChunk == null) {
+						playerDistanceOld = Integer.MAX_VALUE;
+					} else {
+						playerDistanceOld = oldChunk.observers.getLive().get(p);
+						if (playerDistanceOld == null) {
+							playerDistanceOld = Integer.MAX_VALUE;
+						}
+					}
+					Integer playerDistanceNew;
+					if (newChunk == null) {
+						playerDistanceNew = Integer.MAX_VALUE;
+					} else {
+						playerDistanceNew = newChunk.observers.getLive().get(p);
+						if (playerDistanceNew == null) {
+							playerDistanceNew = Integer.MAX_VALUE;
+						}
+					}
+					int entityViewDistanceOld = ((SpoutEntity) e).getPrevViewDistance();
+					int entityViewDistanceNew = e.getViewDistance();
+
+					Player player = (Player) p.getController().getParent();
+
+					if (!player.isOnline()) {
+						continue;
+					}
+					NetworkSynchronizer n = player.getNetworkSynchronizer();
+					if (playerDistanceOld <= entityViewDistanceOld && playerDistanceNew > entityViewDistanceNew) {
+						n.destroyEntity(e);
+					} else if (playerDistanceNew <= entityViewDistanceNew && playerDistanceOld > entityViewDistanceOld) {
+						n.spawnEntity(e);
+					}
+				}
+			}
+		}
+
+		// Update all entities that are in the chunk
+		// TODO - should have sorting based on view distance
+		for (Map.Entry<Entity, Integer> entry : observerLive.entrySet()) {
+			Entity p = entry.getKey();
+			if (p.getController() instanceof PlayerController) {
+				Player player = (Player) p.getController().getParent();
+				if (player.isOnline()) {
+					NetworkSynchronizer n = player.getNetworkSynchronizer();
+					int playerDistance = entry.getValue();
+					Entity playerEntity = p;
+					for (Entity e : entitiesSnapshot) {
+						if (playerEntity != e) {
+							if (playerDistance <= e.getViewDistance()) {
+								if (((SpoutEntity) e).getPrevController() != e.getController()) {
+									n.destroyEntity(e);
+									n.spawnEntity(e);
+								}
+								n.syncEntity(e);
+							}
+						}
+					}
+					for (Entity e : changedEntities) {
+						if (entitiesSnapshot.contains(e)) {
+							continue;
+						} else if (((SpoutEntity) e).justSpawned()) {
+							if (playerEntity != e) {
+								if (playerDistance <= e.getViewDistance()) {
+									if (((SpoutEntity) e).getPrevController() != e.getController()) {
+										n.destroyEntity(e);
+										n.spawnEntity(e);
+									}
+									n.syncEntity(e);
+								}
+							}
+						}
+					}
+				}
+			}
+		}
 	}
 }
